requirements.txt
aniso8601==1.2.1
backports.shutil-get-terminal-size==1.0.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.1.2
enum34==1.1.6
Flask==0.12.2
Flask-RESTful==0.3.6
idna==2.6
ipaddress==1.0.18
ipython==5.4.1
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pymodm==0.4.0
pymongo==3.5.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.2
requests==2.18.4
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
